fix: initPlugins
