+++ packages/builder/builder-rspack-provider/src/core/initPlugins.ts
@@ -58,8 +58,7 @@ export async function initPlugins({
     onAfterBuild: hooks.onAfterBuildHook.tap,
     onBeforeBuild: hooks.onBeforeBuildHook.tap,
     onDevCompileDone: hooks.onDevCompileDoneHook.tap,
-    modifyWebpackChain: hooks.modifyWebpackChainHook.tap,
-    modifyWebpackConfig: hooks.modifyWebpackConfigHook.tap,
+    modifyRspackConfig: hooks.modifyRspackConfigHook.tap,
     modifyBuilderConfig: hooks.modifyBuilderConfigHook.tap,
     onAfterCreateCompiler: hooks.onAfterCreateCompilerHook.tap,
     onBeforeCreateCompiler: hooks.onBeforeCreateCompilerHook.tap,
